test(round-11-agent-07): cover kanban throughput_model + multi_repo
Golden-oracle tests for PoissonRegression.predict (hand-computed values),
predict_probability (exponential CDF), priority encoding, unknown agents,
half-hour floor, monotone probability. Integration tests for ThroughputModel
(predict_completion_time, get_agent_stats with correct keys/values).
Edge cases for extract_card_features (missing FM, bad YAML, empty files,
non-list commits, zero/negative completion times). Tests for multi_repo
(load_cards filtering, generate_multi_repo_status sections, get_cross_repo_status stub).

Bug found:
- BUG-1 (xfail): extract_card_features line 80 — YAML auto-parses ISO
  timestamps as datetime objects, str.replace() fails silently.
  Card frontmatter timestamps must be quoted or code must handle datetime.

Verification:
$ .venv/bin/python3 -m pytest tests/services/kanban/test_throughput_model.py tests/services/kanban/test_multi_repo.py -q
51 passed, 1 xfailed

+++ backend/services/hawkes_process.py
@@ -50,7 +50,7 @@ from typing import Any
 
 logger = logging.getLogger(__name__)
 import numpy as np
-from scipy.optimize import minimize
+from scipy.optimize import minimize  # type: ignore[import-untyped]
 
 logger = logging.getLogger(__name__)
 
@@ -159,7 +159,7 @@ class HawkesProcess:
         else:  # power_law
             excitation = self.alpha * np.sum(1.0 / (dt + self._pl_c) ** (1.0 + self._pl_gamma))
 
-        return self.mu + excitation
+        return float(self.mu + excitation)
 
     def _intensity_vectorized(self, t_array: np.ndarray, event_times: np.ndarray) -> np.ndarray:
         """Vectorized intensity computation for an array of time points."""
@@ -438,7 +438,7 @@ class HawkesProcess:
         t_now = event_times[-1]
         lam = self.intensity(t_now, event_times)
 
-        return 1.0 / max(lam, 1e-300)
+        return float(1.0 / max(lam, 1e-300))
 
     def get_cluster_probability(
         self, event_times: np.ndarray, window: float = 1.0
